[refactor] task 엔터티 생성자 및 id 필드 수정
1. 엔터티 생성자에 빌드 패턴 적용

2. 엔터티 id 필드의 DB 칼럼명 [엔터티명]id로 리팩토링

3. dto 패키지 내부에 request, response 패키지 구조 추가
