+++ easytask/src/main/java/com/easytask/easytask/src/task/entity/TaskUserMapping.java
@@ -2,10 +2,7 @@ package com.easytask.easytask.src.task.entity;
 
 import com.easytask.easytask.common.BaseEntity;
 import com.easytask.easytask.src.user.entity.User;
-import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import javax.persistence.*;
 
@@ -16,7 +13,7 @@ import javax.persistence.*;
 public class TaskUserMapping extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false, updatable = false)
+    @Column(name = "taskUserMappingId", nullable = false, updatable = false)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -26,4 +23,10 @@ public class TaskUserMapping extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "postId")
     private Task task;
+
+    @Builder
+    public TaskUserMapping(User irumi, Task task) {
+        this.irumi = irumi;
+        this.task = task;
+    }
 }
